fix: profile screen rename and imports
