+++ src/screens/Stacks/AllStack.tsx
@@ -1,6 +1,6 @@
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import React from 'react';
-import ProfileScreen2 from '../ProfileScreen/ProfileScreen2';
+import ProfileScreen from '../ProfileScreen/ProfileScreen';
 import TaskDetailScreen from '../ProfileScreen/DetailsScreen/TaskDetailScreen';
 import ProgressDetailScreen from '../ProfileScreen/DetailsScreen/ProgressDetailScreen';
 const Stack = createNativeStackNavigator();
@@ -12,7 +12,7 @@ export function AllTaskScreenStack() {
         headerShown: false,
       }}
     >
-      <Stack.Screen name="Profile" component={ProfileScreen2} />
+      <Stack.Screen name="Profile" component={ProfileScreen} />
       <Stack.Screen name="TaskDetail" component={TaskDetailScreen} />
       <Stack.Screen name="ProgressDetail" component={ProgressDetailScreen} />
     </Stack.Navigator>
